refactor: moved everything from Template to components
