+++ src/Content/Content.test.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import { fireEvent, screen } from '@testing-library/react';
 import Content from './Content';
-import Navigation from '../Template/Navigation/Navigation';
+import Navigation from '../components/Navigation/Navigation';
 import { filmsArray } from '../REDUX/filmsArray';
 import { renderWithRouterAndRedux } from '../App.test';
 import { matchMediaMock } from '../test/matchMediaMock';
