+++ src/Content/Content.tsx
@@ -3,7 +3,7 @@ import { Routes, Route } from 'react-router-dom';
 import About from './About/About';
 import Rules from './Rules/Rules';
 import Seance from './Seance/Seance';
-import Redirect from '../Template/Redirect';
+import Redirect from '../components/Redirect';
 import IndexContent from './IndexContent/IndexContent';
 import SelectedMovieContainer from './Cinema/SelectedMovieContainer';
 const Karaoke = lazy(() => import('./Karaoke/Karaoke'));
