+++ src/Content/IndexContent/IndexContent.tsx
@@ -1,10 +1,9 @@
 import { Link } from 'react-router-dom';
-import { MobileAdv } from '@/Template/Adv';
-import BarSlider from '@/Template/BarSlider';
+import { MobileAdv } from '@/components/Adv';
 import { CINEMA_PHONE } from '@/utils/constants';
 import { convertVideoId } from '@/utils/convertVideoId';
 import { useAppSelector } from '../../REDUX/store';
-import { SushiWork } from '../../Template/SushiWork';
+import { SushiWork } from '../../components/SushiWork';
 
 function IndexContent({ isMobile }: {
   isMobile: boolean;
