+++ src/components/AdvSlider/AdvSlider.css.ts
