+++ src/components/AdvSlider/AdvSlider.tsx
