+++ src/components/AdvSlider/BaseSlide.tsx
