+++ src/components/AdvSlider/Slides.tsx
