+++ src/components/AdvSlider/index.ts
