+++ src/components/AdvSlider/types.ts
