+++ src/components/BarSlider/BarSlider.css.ts
