+++ src/components/BarSlider/BarSlider.tsx
