+++ src/components/BarSlider/index.ts
