+++ src/components/BottomSlider/BottomSlider.css.ts
