+++ src/components/BottomSlider/index.ts
