+++ src/components/FilmImg/index.ts
