+++ src/components/FilmsSlider/FilmsSlider.css.ts
