+++ src/components/FilmsSlider/index.ts
