+++ src/components/Redirect.tsx
