+++ src/components/SpecialSettings/SpecialSettings.tsx
@@ -1,9 +1,9 @@
-import { FontButtons } from './FontButtons';
-import { ImgSwitcher } from './ImgSwitcher';
-import { ThemeButtons } from './ThemeButtons';
 import * as s from './SpecialSettings.css.ts';
-import { SiteModeButton } from './SiteModeButton';
-import { useSpecialState } from '../../REDUX/stateHooks/useSpecialState';
+import { FontButtons } from './FontButtons.tsx';
+import { ImgSwitcher } from './ImgSwitcher.tsx';
+import { ThemeButtons } from './ThemeButtons.tsx';
+import { SiteModeButton } from './SiteModeButton.tsx';
+import { useSpecialState } from '../../REDUX/stateHooks/useSpecialState.ts';
 
 export const SpecialSettings = () => {
   const {
